+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DRmZ3MVp.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-BolErhgA.js";import{_ as a}from"./mermaid.core-BbMA-v1e.js";import"./chunk-5VM5RSS4-B3rZzm1i.js";import"./chunk-XXDRQBXY-cwMt_7c8.js";import"./chunk-POPQ4Y6H-BhIWjX04.js";import"./chunk-F27PBJKO-CQh8xdI8.js";import"./channel-D-8zZ3AR.js";import"./index-q0qmjGGN.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};